ftp://ftp.sunet.se/pub/os/Linux/distributions/ultrapenguin/ultrapenguin-1.1/SRPMS/xscreensaver-2.10-2.src.rpm -rw-r--r-- 1 zblaxell zblaxell 644819 Feb 27 16:56 xscreensaver-2.10-2.src.rpm 6fa56ee5eb79337797c2c9c4c7638a6331703650  xscreensaver-2.10-2.src.rpm
xscreensaver-2.10.tar.gz
-rw-r--r-- 2 zblaxell zblaxell 643509 Oct 23  1997 xscreensaver-2.10.tar.gz
91d4644bc55c8803b7e6741b0748fff3e1013378  xscreensaver-2.10.tar.gz

+++ hacks/qix.c
@@ -18,8 +18,8 @@
 #define SCALE	6
 
 struct qpoint {
-  long x, y;
-  long dx, dy;
+  int x, y;
+  int dx, dy;
 };
 
 struct qline {
